Finished merging
